requirements.txt
# Core Libraries for ETF Trading System
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
jupyter>=1.0.0

# Development Tools
pytest>=7.4.0
ruff>=0.0.280
mypy>=1.5.0

# Flask Web Framework
flask>=2.3.0
flask-sqlalchemy>=3.0.0
python-dotenv>=1.0.0
wtforms>=3.0.0

# Additional Dependencies
requests>=2.31.0
scipy>=1.11.0
scikit-learn>=1.3.0